Score impact of each sample on correlation sturucture

Score impact of each sample on correlation sturucture. Compute correlation using all samples (i.e. C), then compute correlation omitting sample i (i.e. Ci). The score the sample i is based on the difference between C and Ci.

delaneau.score(Y, method = c("pearson", "kendall", "spearman"))

Arguments

Y
data matrix with samples on rows and variables on columns

method
specify which correlation method: "pearson", "kendall" or "spearman"

Value

score for each sample measure impact on correlation structure

See also

Examples

# load iris data
data(iris)

# Evalaute score on each sample
delaneau.score( iris[,1:4] )

#>             1             2             3             4             5 
#>  2.392020e-03 -2.023052e-03 -2.254228e-04 -1.175789e-03  3.067485e-03 
#>             6             7             8             9            10 
#>  3.613378e-03  1.739786e-03  1.610689e-03 -3.581280e-03 -1.058259e-03 
#>            11            12            13            14            15 
#>  3.585132e-03  1.608228e-03 -2.264272e-03 -2.863378e-03  5.269956e-03 
#>            16            17            18            19            20 
#>  4.101427e-03  4.175666e-03  2.310130e-03  3.590239e-03  3.938222e-03 
#>            21            22            23            24            25 
#>  1.847168e-03  3.298703e-03  3.512710e-03  8.622414e-04  1.644556e-03 
#>            26            27            28            29            30 
#> -1.787499e-03  1.532360e-03  2.385847e-03  1.754538e-03 -1.226307e-04 
#>            31            32            33            34            35 
#> -1.039851e-03  1.740320e-03  5.640836e-03  5.352945e-03 -1.002131e-03 
#>            36            37            38            39            40 
#> -2.111842e-05  2.801364e-03  3.216784e-03 -2.461947e-03  1.652132e-03 
#>            41            42            43            44            45 
#>  2.358733e-03 -1.152656e-02 -1.921528e-04  2.156260e-03  3.406659e-03 
#>            46            47            48            49            50 
#> -2.001511e-03  4.051521e-03 -2.039643e-04  3.560882e-03  8.334199e-04 
#>            51            52            53            54            55 
#> -2.621935e-04 -7.436557e-04 -1.444262e-04 -1.576373e-03  8.006617e-04 
#>            56            57            58            59            60 
#>  2.988601e-04 -1.430851e-03 -3.516053e-03  7.904226e-04  1.251058e-04 
#>            61            62            63            64            65 
#> -6.620587e-03  1.136425e-04 -1.713778e-03  3.127337e-04 -4.507567e-05 
#>            66            67            68            69            70 
#>  8.323117e-05  4.918621e-04 -3.178555e-04  6.151750e-04 -1.254188e-03 
#>            71            72            73            74            75 
#> -4.442575e-04  2.056908e-04  1.251319e-03  5.066490e-04  4.963374e-04 
#>            76            77            78            79            80 
#>  3.672432e-04  1.391915e-03 -7.654247e-05  2.766486e-04 -1.282107e-03 
#>            81            82            83            84            85 
#> -1.954137e-03 -2.335353e-03 -3.072779e-04  1.005528e-03  8.558707e-04 
#>            86            87            88            89            90 
#> -1.385069e-03 -2.419081e-04  5.224943e-04  1.949653e-04 -7.713257e-04 
#>            91            92            93            94            95 
#> -2.010578e-04  6.732146e-05 -4.417269e-04 -4.271265e-03 -3.506618e-05 
#>            96            97            98            99           100 
#>  1.276182e-04  1.455780e-04  2.810288e-04 -2.715923e-03  2.715293e-05 
#>           101           102           103           104           105 
#> -2.811279e-03  1.608949e-03 -8.295545e-04  4.658131e-04 -4.556453e-04 
#>           106           107           108           109           110 
#> -9.522773e-04  1.776821e-03  7.473937e-04  2.684696e-03 -9.533994e-03 
#>           111           112           113           114           115 
#> -1.643350e-03  1.352727e-03 -5.915504e-04  2.258187e-03  2.590909e-03 
#>           116           117           118           119           120 
#> -1.672388e-03 -2.684843e-04 -1.375366e-02  3.536663e-03  8.647845e-04 
#>           121           122           123           124           125 
#> -2.868901e-03  1.845266e-03  1.851005e-03  1.123129e-03 -3.362203e-03 
#>           126           127           128           129           130 
#> -2.206105e-03  7.682669e-04  8.593464e-05  1.053592e-03  3.582294e-04 
#>           131           132           133           134           135 
#>  1.653224e-03 -1.216428e-02  1.149715e-03  7.533122e-04  1.591714e-03 
#>           136           137           138           139           140 
#> -9.175706e-04 -3.515552e-03 -8.308748e-04  2.086295e-04 -1.421053e-03 
#>           141           142           143           144           145 
#> -1.533218e-03 -1.239890e-03  1.608949e-03 -2.904417e-03 -3.752880e-03 
#>           146           147           148           149           150 
#> -3.482134e-04  1.903139e-03 -2.864917e-04 -2.970967e-03  3.940631e-04