Plot genes from a specified region of the human genome.

Retrieves genes from the UCSC Genome Browser and generate the genes plot.

plotGenes(minRange, maxRange, chromosome, genome = "hg19", plot_lines_distance = 0.03,
vp = viewport(x = 0, y = 0.99, just = c("left", "top")), splice_variants = TRUE,
non_coding = TRUE)

Arguments

minRange
The sequence minimum range in base pairs.

maxRange
The sequence maximum range in base pairs.

chromosome
A character string identifying the chromosome.

genome
The genome assembly to use. The default is hg19, the most recent human genome assembly on the UCSC genome browser.

plot_lines_distance
The distance between the lines of genes plotted.

vp
A viewport.

splice_variants
If FALSE, exclude gene splice variants.

non_coding
If FALSE, exclude non-coding genes.

Value

A grob of gene plots.

Details

The genes are color coded as follows: Black -- feature has a corresponding entry in the Protein Data Bank (PDB) Dark blue -- transcript has been reviewed or validated by either the RefSeq, SwissProt or CCDS staff Medium blue -- other RefSeq transcripts Light blue -- non-RefSeq transcripts

For assemblies older than hg18, all genes are plotted in grey.

References

http://genome.ucsc.edu/cgi-bin/hgTrackUi?g=knownGene

Author

Sigal Blay <sblay@sfu.ca> and more

Examples

if (FALSE) {
grid.newpage()
plotGenes(149500000, 150000000, "chr7")
}